+++ u64_array_bigints_core/src/parity_uint.rs
@@ -19,12 +19,29 @@ pub use uint::FromStrRadixErr;
 macro_rules! to_from_array {
     ($($to_fn:ident $from_fn:ident $uX:ident $n:expr);*;) => {
         $(
-            pub fn $to_fn(self) -> [$uX; $n] {
-                bytemuck::try_cast(self.0).unwrap()
+            pub fn $from_fn(mut x: [$uX; $n]) -> Self {
+                const_for!(i in {0..x.len()} {
+                    // on little endian this is a no-op, on big endian this fixes the endianness
+                    x[i] = $uX::from_le(x[i]);
+                });
+                let mut res = Self(bytemuck::try_cast(x).unwrap());
+                const_for!(i in {0..res.0.len()} {
+                    // I don't care to think about perf on bigendian, I only care about correctness
+                    res.0[i] = u64::from_le(res.0[i]);
+                });
+                res
             }
 
-            pub fn $from_fn(x: [$uX; $n]) -> Self {
-                Self(bytemuck::try_cast(x).unwrap())
+            pub fn $to_fn(self) -> [$uX; $n] {
+                let mut tmp = self;
+                const_for!(i in {0..tmp.0.len()} {
+                    tmp.0[i] = u64::from_le(self.0[i]);
+                });
+                let mut x: [$uX; $n] = bytemuck::try_cast(tmp.0).unwrap();
+                const_for!(i in {0..x.len()} {
+                    x[i] = $uX::from_le(x[i]);
+                });
+                x
             }
         )*
     };