+++ CHANGELOG.md
@@ -1,3 +1,5 @@
 # Changelog
 
 ## [0.2.0] - TODO
+### Fixes
+- Casting is now correct on big endian systems